SigVerification source code cleanup.

+++ org.jcryptool.visual.sigVerification/src/org/jcryptool/visual/sigVerification/ui/wizards/InputFileComposite.java
@@ -130,6 +130,6 @@ public class InputFileComposite extends Composite implements SelectionListener {
      * @return the value of maxSize in MB (Megabytes)
      */
     public int getMaxSizeInMB() {
-    	return (int) (maxSize / 1048576);
+    	return maxSize / 1048576;
     }
 }
